+++ odoo/local-src/sf_partner_carrier/__manifest__.py
@@ -1,0 +1,17 @@
+# -*- coding: utf-8 -*-
+# Part of SenseFly.
+{
+    "name": "Partner Carrier Account",
+    "version": "10.0.1.0.0",
+    "category": "senseFly",
+    "website": "https://sensefly.com",
+    "author": "Telmo Santos",
+    "application": False,
+    "installable": True,
+    "depends": [
+    ],
+    "data": [
+        'views/carrier_account.xml'
+    ],
+    'images': ['static/description/icon.png'],
+}